Make the top level malaria script simply import and run
Instead of having some of the bootstrap code in place.
This has _zero_ functional change, but gives PyDev a "main module" that
it can use for debugging.  Stupid PyDev. Refused to accept that
"malaria" without a .py could possibly be a python module, even if
though it provides syntax and smart linking on the file.