swarm: support splitting a keyfile and sharing
You can generate a psk file of X pseudo-random key/id pairs with:
$ malaria keygen -n XXXXX > malaria.pskfile

You can divide this file evenly amongst hosts with:
$ fab [vagrant | -H A,B,C,D] share_key:malaria.pskfile

The file will be installed as /tmp/malaria-tmp-keyfile (which you can
override with some fab magic) but that allows it to be used with the
malaria publish command and the "psk_file" option.

Overall, you can now make 10000 keys, copy that keyfile to the server
under test, and then split that keyfile evenly amongst your X attack
nodes.