watch command initial steps
This is the much more passive listener, designed for use with VmstatPlot
for graphing values over time.

The original "listen" command is quite intrusive, it counts messages and
clients and is good for checking for duplicate/missing and so on, but
it's not very good for general status monitoring of the mqtt server.

"watch" just watches a broker, and maintains a pseudo filesystem with
files representing each of the statistics it tracks.