AWS up/down/abort support
Oops, haven't updated the readme yet :(

This improves the way the state file is used, so that you don't need to
fiddle with -H lines as much.  You can now setup the state file with
vagrant up, or beeup:count, and then just use

 fab mstate attack:blah

There's still niggles and wrinkles with making AWS attack nodes the same
as any other sort of node, but it's largely cleaner.